refactor games to receive data as props
